Optimize removeSeqPerSubject() for MaxMsgPerSubject == 1 (#6688)
The following PR builds on top of PR
https://github.com/nats-io/nats-server/pull/6686 by @svenfoo:

> This brings back an optimization that was removed with commit
79b41b122511df.
>
> Without this optimization the first sequence number is lost from the
per-subject state whenever a message is added and the per subject limit
is enforced. This is happening a lot on a stream with MaxMsgPerSubject
set to 1. And it can lead to a dramatic slowdown when catching up on a
stream with sequence numbers spread out wide.

Resolves https://github.com/nats-io/nats-server/issues/6687

Signed-off-by: Sven Neumann <sven.neumann@holoplot.com>
Signed-off-by: Maurice van Veen <github@mauricevanveen.com>